refactor(agent): split dispatcher command handlers
